requirements.txt
numpy
opencv-python
matplotlib
h5py
torch
torchvision
pyyaml